backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-business-flow-backend"
version = "0.1.0"
description = "Backend API for AI Business Flow application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "pymysql==1.1.0",
    "cryptography==41.0.7",
    "python-dotenv==1.0.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
    "openai==1.3.7",
    "python-multipart==0.0.6",
    "alembic==1.13.1",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
    "pytest-env==1.1.3",
    "httpx==0.25.2",
    "hypothesis==6.92.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
env = [
    "PYTEST_CURRENT_TEST=1"
]